requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
praw==5.4.0
prawcore==0.14.0
pycryptodome==3.6.3
requests==2.19.1
update-checker==0.16
urllib3==1.23
